requirements.txt
Django
django-rest-framework
pytube==12.1.0
Pillow==9.2.0
django-cors-headers==3.13.0
requests==2.28.1
djangorestframework-simplejwt==5.2.2
pyodbc==4.0.34
mssql-django==1.1.3
django-stdimage==6.0.1
python-decouple==3.6